scripts/requirements.txt
pandas
thefuzz[speedup]
anthropic
requests
tqdm
python-dotenv
